extract upload code into TS mixin
